+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Code comment after expression is assigned to parenthesis. [#1864](https://github.com/fsprojects/fantomas/issues/1864)
 - Misaligned comment in MultiLineBlockBracket. [#1716](https://github.com/fsprojects/fantomas/issues/1716)
 - Comment unindented after DU cases. [#2606](https://github.com/fsprojects/fantomas/issues/2606)
+- Comment on last match case has wrong indentation. [#2653](https://github.com/fsprojects/fantomas/issues/2653)
 
 ## [8.0.0-alpha-008] - 2026-03-25